layout/generic/test/test_contained_plugin_transplant.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=775965
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 775965</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body onload="run();">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=775965">Mozilla Bug 775965</a>
<p id="display"></p>
<div id="content">

<iframe id="iframe1" src="data:text/html,<embed type='application/x-test' width='200' height='200'></embed>"></iframe>
<iframe id="iframe2"></iframe>

</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 775965 **/
SimpleTest.waitForExplicitFinish();

function run() {
  var iframe1 = document.getElementById("iframe1");
  var iframe2 = document.getElementById("iframe2");
  iframe1.parentNode.removeChild(iframe1);
  iframe1.style.position = "fixed";
  iframe2.contentDocument.body.appendChild(iframe1);
  setTimeout(function() { ok(true, "We didn't crash!"); SimpleTest.finish(); }, 0);
}


</script>
</pre>
</body>
</html>
